Перезалив
Полностью переписал код с использованием предыдущих рабочих функций, более плотно и модульно структурировал файлы.
